// packages/ew-utils-deployment/src/types/types.d.ts
export declare interface Tx {
  nonce?: string | number
  chainId?: string | number
  from?: string
  to?: string
  data?: string
  value?: string | number
  gas?: string | number
  gasPrice?: string | number,
  privateKey?: string
}